pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-helpers"
version = "0.1.7"
description = "Python helpers for the workspace"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.39.14",
    "botocore>=1.39.14",
    "casbin>=1.43.0",
    "dotenv>=0.9.9",
    "email-validator>=2.0.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "moto>=5.1.8",
    "pycognito>=2024.5.1",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "setuptools>=80.9.0",
    "starlette>=0.47.2",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "moto>=5.0.0",
    "httpx>=0.25.0",
]

[tool.setuptools.packages.find]
include = ["custom_auth", "custom_aws", "custom_cognito"]
exclude = ["tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "integration_real: Real integration tests with AWS services",
    "slow: Slow tests",
]
